Organize Group Element Tests
